Update example
Issue: VTBDBOTND-408

+++ carrent-carmanagement-impl/src/main/java/ru/vtb/carrent/car/domain/entity/Car.java
@@ -50,7 +50,7 @@ public class Car {
     /**
      * Пробег на последнюю дату в статусе в наличии
      */
-    private double mileage;
+    private int mileage;
 
     /**
      * Дата последнего ТО
